pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
dynamic = ["version"]
name = "nu_bot"
readme = "README.md"
dependencies = [
    "pandas",
    "numpy",
    "flask",
    "gunicorn",
    "flask_restx",
    "python-dotenv",
    "requests",
    "beautifulsoup4",
    "datasets",
    "transformers",
    "torch",
    "faiss-cpu",
    "prefect",
    "streamlit",
    "sentence-transformers",
    "dvc-gs"
]
requires-python = ">= 3.8"
keywords = ["nu_bot","Nubot"]

[project.urls]
Homepage="https://github.com/udishadc/NUBot"
Repository="https://github.com/udishadc/NUBot"
